Add icon.png and app.json files to javascript folders to support Intel XDK IoT samples.
These two files are not essential to the app. The app.json file contains metadata that is used by the Intel XDK "feed" system. The samples UI also uses data in that file to display information about the samples. The icon.png file is used by the XDK samples UI to display a unique image for the IoT sample. Each of these files (app.json and icon.png) are unique to the individual sample.